REF: Move test to better location
